Added the Example flower asset in unity
Along with the shader. I also set up the material but it didn't work with my shader and I'm too scared to touch it in any way, so I just left it with the settings I used in my example project.
One difference I noticed from my own project is that this project is using the pbr template, while the one I used was just the standard 3D one.